separate types into different files
