+++ client-v2/src/actions/FrontsConfig.js
@@ -2,7 +2,7 @@
 
 import type { Action } from '../types/Action';
 import type { ThunkAction } from '../types/Store';
-import type { FrontsConfig } from '../types/Fronts';
+import type { FrontsConfig } from '../types/FrontsConfig';
 
 import { fetchFrontsConfig } from '../services/faciaApi';
