+++ client-v2/src/selectors/__tests__/frontsSelectors.spec.js
@@ -2,7 +2,7 @@
 
 import { getFrontsConfig } from '../frontsSelectors';
 import { frontsConfig } from '../../fixtures';
-import type { FrontDetail } from '../../types/Fronts';
+import type { FrontDetail } from '../../types/FrontsConfig';
 
 const editorialFronts: Array<FrontDetail> = [
   { collections: ['collection1'], id: 'editorialFront' }
