+++ client-v2/src/types/Article.js
@@ -1,0 +1,13 @@
+// @flow
+
+// TODO: fill this in as we start to display overrides
+type Meta = Object;
+
+type Article = {
+  id: string,
+  frontPublicationDate: number,
+  publishedBy: string,
+  meta: Meta
+};
+
+export type { Article };
