+++ client-v2/src/types/Capi.js
@@ -1,0 +1,7 @@
+// @flow
+
+type CapiArticle = {
+  headline: string
+};
+
+export type { CapiArticle };
