+++ client-v2/src/types/Priority.js
@@ -1,0 +1,12 @@
+// @flow
+
+type PriorityName = 'editorial' | 'commercial' | 'training' | 'email';
+
+type Priorities = {
+  editorial: Object,
+  commercial: Object,
+  training: Object,
+  email: Object
+};
+
+export type { PriorityName, Priorities };
